+++ src/main/java/co/uk/jiveelection/campaign/App.java
@@ -1,6 +1,6 @@
 package co.uk.jiveelection.campaign;
 
-import co.uk.jiveelection.campaign.jive.InMemoryJiveTranslator;
+import co.uk.jiveelection.campaign.jive.memory.InMemoryJiveTranslator;
 import co.uk.jiveelection.campaign.twit.TweetHelper;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
